+++ .gitignore
@@ -1,4 +1,4 @@
 **/.DS_Store
 clouds
 .ipynb_checkpoints
-__pycache__
+**/__pycache__
