+++ dataset_creation/kb_crawl/classes/property.py
@@ -1,0 +1,5 @@
+class Property:
+  def __init__(self, original: str, antonym: [str] = [], paraphrase: [str] = []):
+    self.original = original
+    self.antonym = antonym
+    self.paraphrase = paraphrase
